+++ ui-spa/src/components/egress-connect/EgressSearchPage.tsx
@@ -1,7 +1,7 @@
 import { useRef, useEffect } from "react";
 import { Button, Input, InsetText, ErrorSummary, BackLink } from "../govuk";
 import EgressSearchResults from "./EgressSearchResults";
-import { UseApiResult } from "../../common/hooks/useApiNew";
+import { UseApiResult } from "../../common/hooks/useApi";
 import { EgressSearchResultData } from "../../common/types/EgressSearchResponse";
 import styles from "./EgressSearchPage.module.scss";
